requirements.txt
flask==2.3.3
gunicorn==21.2.0
pytest==7.4.0
pytest-cov==4.1.0
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0
pyYAML==6.0.1
requests==2.31.0
tinydb==4.8.0
click==8.1.7
tqdm==4.66.1
rich==13.7.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
prometheus-client==0.17.1
structlog==23.1.0
httpx==0.24.1